+++ .cargo/config.toml
@@ -7,7 +7,6 @@
 [profile.dev]
 opt-level = 1
 
-# Enable high optimizations for dependencies (incl. Bevy), but not for our code:
 [profile.dev.package."*"]
 opt-level = 2 # lowered opt level to avoid crash in macos - see https://github.com/bevyengine/bevy/issues/10524
