+++ bevy_dexterous_developer/src/hot_internal/hot_reload_internal.rs
@@ -1,7 +1,8 @@
 use bevy::{prelude::*, window::PrimaryWindow};
 
-pub use crate::hot_internal::hot_reload_internal::InternalHotReload;
 use crate::{ReloadMode, ReloadSettings};
+
+use super::reload_systems::InternalHotReload;
 
 pub fn draw_internal_hot_reload(
     internal: Res<InternalHotReload>,
@@ -25,7 +26,7 @@ pub fn draw_internal_hot_reload(
         .unwrap_or("all reloadables")
         .replace("_dexterous_developered_inner_", "_");
 
-    let update = internal
+    let update = internal.0
         .last_update_date_time
         .format("%H:%M:%S")
         .to_string();
