+++ dexterous_developer/src/lib.rs
@@ -1,5 +1,8 @@
 pub use dexterous_developer_internal::*;
 
+#[cfg(feature = "bevy")]
+pub use bevy_dexterous_developer::*;
+
 #[cfg(feature = "hot_internal")]
 #[allow(unused_imports)]
 #[allow(clippy::single_component_path_imports)]
