+++ dexterous_developer_cli/Cargo.toml
@@ -11,7 +11,7 @@ version.workspace = true
 edition = "2021"
 readme = "../README.md"
 exclude = ["assets/**/*", "tools/**/*", ".github/**/*"]
-description = "A CLI for hot reloadable apps in bevy, using the dexterous_developer crate"
+description = "A CLI for hot reloadable Rust applications, using the dexterous_developer crate"
 
 [lib]
 crate-type = ["rlib"]
