+++ dexterous_developer_dynamic/Cargo.toml
@@ -16,9 +16,12 @@ version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
 
+bevy = ["dep:bevy_dexterous_developer"]
+
 [lib]
 crate-type = ["dylib"]
 
 
 [dependencies]
 dexterous_developer_internal = { version = "0.1.0", path = "../dexterous_developer_internal", default-features = false }
+bevy_dexterous_developer = { version = "0.1.0", path = "../bevy_dexterous_developer", default-features = false, optional = true}
