+++ dexterous_developer_dynamic/src/lib.rs
@@ -1,1 +1,2 @@
 pub use dexterous_developer_internal::*;
+pub use bevy_dexterous_developer::*;
