+++ dexterous_developer_internal/src/lib.rs
@@ -1,5 +1,3 @@
-#[cfg(all(feature = "hot_internal", feature = "bevy"))]
-mod hot_internal;
 
 #[cfg(feature = "hot")]
 mod hot;
@@ -10,10 +8,11 @@ mod cold;
 #[cfg(any(feature = "hot", feature = "hot_internal"))]
 pub mod internal_shared;
 
+#[cfg(feature = "hot_internal")]
+pub mod hot_internal;
+
 mod types;
 
-#[cfg(feature = "bevy")]
-pub mod bevy_support;
 mod logger;
 
 pub use dexterous_developer_macros::*;
@@ -28,5 +27,3 @@ pub use hot::{
     compile_reloadable_libraries, run_existing_library, run_served_file, watch_reloadable,
 };
 
-#[cfg(feature = "bevy")]
-pub use bevy_support::*;
