+++ testing/templates/no_cli_test/simple/Cargo.toml
@@ -11,4 +11,4 @@ crate-type = ["rlib", "dylib"]
 
 [dependencies]
 bevy = { version = "0.12", default_features = false }
-dexterous_developer = { path = "../../../../dexterous_developer" }
+dexterous_developer = { path = "../../../../dexterous_developer", features = ["bevy"] }
