+++ testing/templates/reloadables_test/Cargo.toml
@@ -10,7 +10,7 @@ crate-type = ["rlib", "dylib"]
 
 [dependencies]
 bevy = { version = "0.12", default_features = false }
-dexterous_developer = { path = "../../../dexterous_developer" }
+dexterous_developer = { path = "../../../dexterous_developer", features = ["bevy"] }
 serde = { version = "1", features = ["derive"] }
 
 [workspace]
