+++ testing/templates/simple_cli_test/Cargo.toml
@@ -11,6 +11,6 @@ crate-type = ["rlib", "dylib"]
 
 [dependencies]
 bevy = { version = "0.12", default_features = false }
-dexterous_developer = { path = "../../../dexterous_developer" }
+dexterous_developer = { path = "../../../dexterous_developer", features = ["bevy"] }
 
 [workspace]
